[core] Scale YCSB to 2^63 operations and/or records.
Updates the generators and measurements classes to use longs for ids and
counts. Also enforced where appropriate that the generated values will
not become negative. Instead the value will stop at Long.MAX_VALUE
avoiding issues for distributions that are looking for values in a
defined range.

Fixes #384

+++ core/src/main/java/com/yahoo/ycsb/Client.java
@@ -237,10 +237,10 @@ class ClientThread extends Thread
   DB _db;
   boolean _dotransactions;
   Workload _workload;
-  int _opcount;
+  long _opcount;
   double _targetOpsPerMs;
 
-  int _opsdone;
+  long _opsdone;
   int _threadid;
   int _threadcount;
   Object _workloadstate;
@@ -259,7 +259,7 @@ class ClientThread extends Thread
    * @param targetperthreadperms target number of operations per thread per ms
    * @param completeLatch The latch tracking the completion of all clients.
    */
-  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, int opcount, double targetperthreadperms, CountDownLatch completeLatch)
+  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, long opcount, double targetperthreadperms, CountDownLatch completeLatch)
   {
     _db=db;
     _dotransactions=dotransactions;
@@ -276,7 +276,7 @@ class ClientThread extends Thread
     _completeLatch=completeLatch;
   }
 
-  public int getOpsDone()
+  public long getOpsDone()
   {
     return _opsdone;
   }
@@ -400,9 +400,9 @@ class ClientThread extends Thread
   /**
    * the total amount of work this thread is still expected to do
    */
-  public int getOpsTodo()
+  public long getOpsTodo()
   {
-    int todo = _opcount - _opsdone;
+    long todo = _opcount - _opsdone;
     return todo < 0 ? 0 : todo;
   }
 }
@@ -515,7 +515,7 @@ public class Client
    * loaded from conf.
    * @throws IOException Either failed to write to output stream or failed to close it.
    */
-  private static void exportMeasurements(Properties props, int opcount, long runtime)
+  private static void exportMeasurements(Properties props, long opcount, long runtime)
       throws IOException
   {
     MeasurementsExporter exporter = null;
@@ -818,20 +818,20 @@ public class Client
 
     System.err.println("Starting test.");
 
-    int opcount;
+    long opcount;
     if (dotransactions)
     {
-      opcount=Integer.parseInt(props.getProperty(OPERATION_COUNT_PROPERTY,"0"));
+      opcount=Long.parseLong(props.getProperty(OPERATION_COUNT_PROPERTY,"0"));
     }
     else
     {
       if (props.containsKey(INSERT_COUNT_PROPERTY))
       {
-        opcount=Integer.parseInt(props.getProperty(INSERT_COUNT_PROPERTY,"0"));
+        opcount=Long.parseLong(props.getProperty(INSERT_COUNT_PROPERTY,"0"));
       }
       else
       {
-        opcount=Integer.parseInt(props.getProperty(RECORD_COUNT_PROPERTY, DEFAULT_RECORD_COUNT));
+        opcount=Long.parseLong(props.getProperty(RECORD_COUNT_PROPERTY, DEFAULT_RECORD_COUNT));
       }
     }
 
@@ -851,7 +851,7 @@ public class Client
       }
 
 
-      int threadopcount = opcount/threadcount;
+      long threadopcount = opcount/threadcount;
 
       // ensure correct number of operations, in case opcount is not a multiple of threadcount
       if (threadid<opcount%threadcount)
@@ -892,7 +892,7 @@ public class Client
       terminator.start();
     }
 
-    int opsDone = 0;
+    long opsDone = 0;
 
     for (Thread t : clients)
     {
