+++ core/src/test/java/com/yahoo/ycsb/measurements/exporter/TestMeasurementsExporter.java
@@ -42,8 +42,8 @@ public class TestMeasurementsExporter {
         long max = 100000;
         ZipfianGenerator zipfian = new ZipfianGenerator(min, max);
         for (int i = 0; i < 1000; i++) {
-            int rnd = zipfian.nextInt();
-            mm.measure("UPDATE", rnd);
+            long rnd = zipfian.nextLong();
+            mm.measure("UPDATE", (int) rnd);
         }
         mm.exportMeasurements(export);
         export.close();
